feat: extract the GitHubLink feature from App to a dedicated component
